✏️ [Modify] TabBar Helper 메소드 객체화
