+++ Sodam/Sodam/View/Storage/HangdamStorageView.swift
@@ -33,26 +33,10 @@ struct HangdamStorageView: View {
             .padding(.horizontal)
             .background(Color.viewBackground)
             .onAppear {
-                withAnimation {
-                    if let tabBarController = getRootTabBarController() {
-                        tabBarController.tabBar.isHidden = false
-                    }
-                }
                 viewModel.loadHangdams()
             }
         }
         .tint(.textAccent)
-    }
-    
-    // MARK: - Helper method
-    
-    private func getRootTabBarController() -> UITabBarController? {
-        guard let scene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
-              let sceneDelegate = scene.delegate as? SceneDelegate,
-              let rootViewController = sceneDelegate.window?.rootViewController else {
-            return nil
-        }
-        return rootViewController as? UITabBarController
     }
 }
